:broken_heart: Deleting WaitFor... going better
When first creating a PR to Hamcrest for the WaitFor matcher it was remarked that it does not match Hamcrest's requirements of matchers being functional only.

Trying to incorporate this feedback a new wait-approach evolved - which now even works better than the original WaitFor matcher, with a much more flexible API - and sticking to the functional requirements of Hamcrest.